# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Meat N Sea API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
  /** @nullable */
  devOtp?: string | null;
}

export interface AuthToken {
  message: string;
  userId: string;
  role: string;
  token: string;
}

export interface OtpRequest {
  phone: string;
}

export interface OtpVerify {
  phone: string;
  code: string;
}

export interface Vendor {
  _id: string;
  name: string;
  phone: string;
  isOpen: boolean;
  status: string;
  serviceRadiusKm: number;
  /** @nullable */
  fssaiNumber?: string | null;
  /** @nullable */
  distance?: number | null;
}

export interface VendorStatusInput {
  isOpen: boolean;
}

export interface Product {
  _id: string;
  vendorId: string;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  category?: string | null;
  pricePaise: number;
  stockQuantity: number;
  isOutOfStock: boolean;
  /** @nullable */
  imageUrl?: string | null;
}

export interface StockToggleInput {
  /** @nullable */
  stockQuantity?: number | null;
}

export interface Order {
  _id: string;
  customerId: string;
  vendorId: string;
  /** @nullable */
  partnerId?: string | null;
  totalAmountPaise: number;
  paymentMethod: string;
  currentStatus: string;
  /** @nullable */
  customerNote?: string | null;
  createdAt?: string;
}

export type OrderInputUserLocation = {
  lng: number;
  lat: number;
};

export interface OrderInput {
  vendorId: string;
  userLocation: OrderInputUserLocation;
  /** @nullable */
  customerNote?: string | null;
  paymentMethod?: string;
}

export interface OrderStatusInput {
  newStatus: string;
}

export interface WhatsappLink {
  link: string;
}

export type DailyReportStatusBreakdownItem = {
  _id?: string;
  count?: number;
};

export interface DailyReport {
  totalOrders: number;
  grossRevenuePaise: number;
  platformFeePaise: number;
  /** @nullable */
  topVendorId?: string | null;
  /** @nullable */
  topVendorName?: string | null;
  totalUsers?: number;
  totalVendors?: number;
  statusBreakdown?: DailyReportStatusBreakdownItem[];
}

export interface SearchResult {
  vendors: Vendor[];
  products: Product[];
  query: string;
}

export interface StudioPlan {
  _id: string;
  name: string;
  description: string;
  pricePaise: number;
  intervalDays: number;
  curatedItems?: string[];
  isActive?: boolean;
}

export interface StudioPlanInput {
  name: string;
  description: string;
  pricePaise: number;
  intervalDays: number;
  curatedItems?: string[];
}

export interface Collection {
  _id: string;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  freshnessHours?: number | null;
  /** @nullable */
  harvestTime?: string | null;
  isActive?: boolean;
}

export interface StudioHome {
  collections: Collection[];
  plans: StudioPlan[];
  featuredProducts: Product[];
}

export type AnalyticsSummaryEventCountsItem = {
  _id?: string;
  count?: number;
};

export type AnalyticsSummaryOrderTrendItem = {
  _id?: string;
  orders?: number;
  revenuePaise?: number;
};

export interface AnalyticsSummary {
  eventCounts: AnalyticsSummaryEventCountsItem[];
  orderTrend: AnalyticsSummaryOrderTrendItem[];
}

export type GetAllOrdersParams = {
status?: string;
limit?: string;
};

export type GetNearbyVendorsParams = {
lng: string;
lat: string;
maxDistance?: string;
};

export type GetProductsParams = {
vendorId?: string;
category?: string;
};

export type GlobalSearchParams = {
q: string;
limit?: string;
};